extracted platform info drawer to a component
